+++ src/test/java/org/openstreetmap/atlas/geography/atlas/items/complex/aoi/ComplexAOIRelationFinderTestRule.java
@@ -1,4 +1,4 @@
-package org.openstreetmap.atlas.geography.atlas.items.complex.relation;
+package org.openstreetmap.atlas.geography.atlas.items.complex.aoi;
 
 import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.tags.RelationTypeTag;
